DF-10
* Refactor Heartbeat and Message to be abstract
* Use polymorphism to resolve concrete class type
* Remove abstract class instance-specific tests
