+++ src/main/java/messaging/HeartbeatMajor.java
@@ -1,15 +1,26 @@
 package messaging;
 
+import util.Host;
+
 import java.net.UnknownHostException;
 
 public class HeartbeatMajor extends Heartbeat {
 
     public HeartbeatMajor(Integer totalChunksMaintained, Long freeSpaceAvailable) throws UnknownHostException {
-        super(MessageType.HEARTBEAT_MAJOR, totalChunksMaintained, freeSpaceAvailable);
+        this(Host.getHostname(), Host.getIpAddress(), 9001, totalChunksMaintained, freeSpaceAvailable);
     }
 
     public HeartbeatMajor(String hostname, String ipAddress, Integer port, Integer totalChunksMaintained,
                           Long freeSpaceAvailable) {
-        super(MessageType.HEARTBEAT_MAJOR, hostname, ipAddress, port, totalChunksMaintained, freeSpaceAvailable);
+        this.hostname = hostname;
+        this.ipAddress = ipAddress;
+        this.port = port;
+        this.totalChunksMaintained = totalChunksMaintained;
+        this.freeSpaceAvailable = freeSpaceAvailable;
+    }
+
+    @Override
+    public MessageType getType() {
+        return MessageType.HEARTBEAT_MAJOR;
     }
 }
